backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
celery[redis]==5.4.0
redis==5.2.1
httpx==0.28.1
minio==7.2.12
cryptography==44.0.0
websockets==14.1
python-dotenv==1.0.1
